Restrucutred Autonomous, cleaned up file structure
