+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Hardware.java
@@ -1,21 +1,10 @@
 package org.firstinspires.ftc.teamcode;
 
-import androidx.annotation.NonNull;
-
-import com.qualcomm.hardware.bosch.BHI260IMU;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotor.RunMode;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.IMU;
-import com.qualcomm.robotcore.hardware.MotorControlAlgorithm;
-import com.qualcomm.robotcore.hardware.PIDFCoefficients;
 import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.RobotLog;
 
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.openftc.easyopencv.OpenCvCamera;
 
 public abstract class Hardware extends LinearOpMode {
@@ -30,7 +19,6 @@ public abstract class Hardware extends LinearOpMode {
     public Servo drone;
 
     public OpenCvCamera camera;
-    public AprilTagDetectionPipeline aprilTagDetectionPipeline;
     public void initHardware() {
         // Initialize motors
         leftFrontDrive = hardwareMap.get(DcMotorEx.class, "FLdrive");
