+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOpTesting.java
@@ -1,14 +1,14 @@
 package org.firstinspires.ftc.teamcode;
 
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
+
+import org.firstinspires.ftc.robotcore.external.Telemetry;
 
 @TeleOp(name = "MotorVerification")
 @Disabled
