+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/AutoBF_Full.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autonomous;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
@@ -8,15 +8,16 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
+
 import org.firstinspires.ftc.robotcore.external.Telemetry;
+import org.firstinspires.ftc.teamcode.MecanumTrain;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "Autonomous BlueBack With Camera")
-public class AutoBB_WithCamera extends LinearOpMode {
+@Autonomous(name = "Autonomous BlueFront Full")
+public class AutoBF_Full extends LinearOpMode {
     MecanumTrain bot;
-
-    enum SPIKE_LOC  {
+    enum SPIKE_LOC {
         LEFT,
         CENTER,
         RIGHT,
@@ -24,107 +25,118 @@ public class AutoBB_WithCamera extends LinearOpMode {
     }
 
     @Override
-    public void runOpMode() throws InterruptedException {
+    public void runOpMode() throws InterruptedException{
         Telemetry telemetry = new MultipleTelemetry(this.telemetry, FtcDashboard.getInstance().getTelemetry());
         ElapsedTime runtime = new ElapsedTime();
         bot = new MecanumTrain(hardwareMap, runtime);
 
-        Pose2d startPose = new Pose2d(-35, 59, Math.toRadians(0));
+        Pose2d startPose = new Pose2d(10.5, 59, Math.toRadians(0));
         bot.odometry.setPoseEstimate(startPose);
 
-        
-        Trajectory findBlueBProp = bot.odometry.trajectoryBuilder(startPose)
-                .lineTo(new Vector2d(-35, 53))
+        Trajectory findBlueFProp = bot.odometry.trajectoryBuilder(startPose)
+                .lineTo(new Vector2d(12, 53))
                 .build();
 
-        // TODO: Blue-Back Drive Trajectories
-        TrajectorySequence BBdriveToLSpike = bot.odometry.trajectorySequenceBuilder(findBlueBProp.end())
-                .splineToSplineHeading(new Pose2d(-46, 35, Math.toRadians(270)), Math.toRadians(270))
+        TrajectorySequence BFdriveToLSpike = bot.odometry.trajectorySequenceBuilder(findBlueFProp.end())
+                .splineToSplineHeading(new Pose2d(23, 38, Math.toRadians(270)), Math.toRadians(0))
                 .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.openClaw())
-                .back(6)
+                .back(3)
+                .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.closeClaw())
                 .waitSeconds(0.5)
-                .strafeRight(3)
-                .splineToSplineHeading(new Pose2d(-47, 9.5, Math.toRadians(90)), Math.toRadians(0))
-                .splineToConstantHeading(new Vector2d(5, 13), Math.toRadians(0))
-                .UNSTABLE_addDisplacementMarkerOffset(24, () -> bot.target = 130)
-                .splineToSplineHeading(new Pose2d(52, 33, Math.toRadians(180)), Math.toRadians(0))
+                .back(9)
+                .UNSTABLE_addDisplacementMarkerOffset(6, () -> {bot.target = 110;})
                 .setAccelConstraint(bot.odometry.SHAKE_ACCEL_CONSTRAINT)
-                .forward(6)
-                .back(6)
+                .forward(4)
+                .back(4)
                 .resetAccelConstraint()
+                .lineTo(new Vector2d(25, 43))
+                .splineToSplineHeading(new Pose2d(53, 35, Math.toRadians(180)), Math.toRadians(0))
                 .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.openClaw())
                 .waitSeconds(0.5)
-                .addTemporalMarker(() -> bot.target = 80)
+                .addTemporalMarker(() -> {bot.target = 80;})
                 .addTemporalMarker(() -> bot.closeClaw())
-                .waitSeconds(1)
-                .lineTo(new Vector2d(36, 45))
-                .addTemporalMarker(() -> bot.target = 40)
-                .waitSeconds(0.4)
-                .addTemporalMarker(() -> bot.target = 0)
-                .splineToConstantHeading(new Vector2d(60, 60), Math.toRadians(0))
+                .waitSeconds(0.5)
+                .lineTo(new Vector2d(45, 15))
+                .addTemporalMarker(() -> {bot.target = 40;})
+                .waitSeconds(0.3)
+                .addTemporalMarker(() -> {bot.target = 0;})
+                .lineTo(new Vector2d(60, 10))
                 .build();
 
-        TrajectorySequence BBdriveToCSpike = bot.odometry.trajectorySequenceBuilder(findBlueBProp.end())
-                .splineToSplineHeading(new Pose2d(-41, 24, Math.toRadians(315)), Math.toRadians(315))
+        TrajectorySequence BFdriveToCSpike = bot.odometry.trajectorySequenceBuilder(findBlueFProp.end())
+                .lineTo(new Vector2d(14, 40))
+                .splineToSplineHeading(new Pose2d(18, 24, Math.toRadians(230)), Math.toRadians(270))
                 .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.openClaw())
-                .back(6)
+                .back(4)
                 .addTemporalMarker(() -> bot.closeClaw())
                 .waitSeconds(0.5)
-                .strafeRight(3)
-                .splineToSplineHeading(new Pose2d(-47, 9.5, Math.toRadians(90)), Math.toRadians(0))
-                .splineToConstantHeading(new Vector2d(5, 13), Math.toRadians(0))
-                .UNSTABLE_addDisplacementMarkerOffset(24, () -> bot.target = 130)
-                .splineToSplineHeading(new Pose2d(52, 33, Math.toRadians(180)), Math.toRadians(0))
+                .back(6)
+                .strafeLeft(3)
                 .setAccelConstraint(bot.odometry.SHAKE_ACCEL_CONSTRAINT)
-                .forward(6)
-                .back(6)
+                .forward(4)
+                .back(4)
                 .resetAccelConstraint()
+                .UNSTABLE_addDisplacementMarkerOffset(10, () -> {
+                    bot.target = 110;
+                })
+                .lineToSplineHeading(new Pose2d(53, 31, Math.toRadians(180)))
                 .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.openClaw())
                 .waitSeconds(0.5)
-                .addTemporalMarker(() -> bot.target = 80)
+                .addTemporalMarker(() -> {
+                    bot.target = 80;
+                })
                 .addTemporalMarker(() -> bot.closeClaw())
-                .waitSeconds(1)
-                .lineTo(new Vector2d(36, 45))
-                .addTemporalMarker(() -> bot.target = 40)
-                .waitSeconds(0.4)
-                .addTemporalMarker(() -> bot.target = 0)
-                .splineToConstantHeading(new Vector2d(60, 60), Math.toRadians(0))
+                .waitSeconds(0.5)
+                .lineTo(new Vector2d(42, 10))
+                .addTemporalMarker(() -> {
+                    bot.target = 40;
+                })
+                .waitSeconds(0.2)
+                .addTemporalMarker(() -> {
+                    bot.target = 0;
+                })
+                .lineTo(new Vector2d(60, 8))
                 .build();
 
-        TrajectorySequence BBdriveToRSpike = bot.odometry.trajectorySequenceBuilder(findBlueBProp.end())
-                .lineToSplineHeading(new Pose2d(-38, 40, Math.toRadians(0)))
-                .splineToConstantHeading(new Vector2d(-28, 34), Math.toRadians(180))
+        TrajectorySequence BFdriveToRSpike = bot.odometry.trajectorySequenceBuilder(findBlueFProp.end())
+                .lineToSplineHeading(new Pose2d(15, 45, Math.toRadians(180)))
+                .splineToConstantHeading(new Vector2d(4, 29), Math.toRadians(180))
+                .waitSeconds(.5)
                 .addTemporalMarker(() -> bot.openClaw())
-                .waitSeconds(0.5)
                 .back(5)
-                .waitSeconds(0.5)
                 .addTemporalMarker(() -> bot.closeClaw())
                 .waitSeconds(0.5)
-                .strafeRight(5)
-                .splineToConstantHeading(new Vector2d(-27, 11), Math.toRadians(0))
-                .UNSTABLE_addDisplacementMarkerOffset(48, () -> bot.target = 130)
-                .lineToConstantHeading(new Vector2d(10, 11))
-                .splineToSplineHeading(new Pose2d(52, 39, Math.toRadians(180)), Math.toRadians(0))
+                .back(8)
                 .setAccelConstraint(bot.odometry.SHAKE_ACCEL_CONSTRAINT)
                 .forward(6)
                 .back(6)
                 .resetAccelConstraint()
-                .waitSeconds(0.5)
+                .UNSTABLE_addDisplacementMarkerOffset(12, () -> {
+                    bot.target = 110;
+                })
+                .lineToConstantHeading(new Vector2d(53, 27))
+                .waitSeconds(.5)
                 .addTemporalMarker(() -> bot.openClaw())
-                .waitSeconds(0.5)
-                .addTemporalMarker(() -> bot.target = 80)
+                .waitSeconds(.8)
+                .addTemporalMarker(() -> {
+                    bot.target = 80;
+                })
                 .addTemporalMarker(() -> bot.closeClaw())
                 .waitSeconds(1)
-                .lineTo(new Vector2d(36,45))
-                .addTemporalMarker(() -> bot.target = 40)
-                .waitSeconds(0.4)
-                .addTemporalMarker(() -> bot.target = 0)
-                .splineToConstantHeading(new Vector2d(60, 60), Math.toRadians(0))
+                .lineTo(new Vector2d(42, 13))
+                .addTemporalMarker(() -> {
+                    bot.target = 40;
+                })
+                .waitSeconds(.4)
+                .addTemporalMarker(() -> {
+                    bot.target = 0;
+                })
+                .lineTo(new Vector2d(60, 8))
                 .build();
 
         bot.initEocvBlue(hardwareMap);
@@ -137,8 +149,9 @@ public class AutoBB_WithCamera extends LinearOpMode {
         if (opModeIsActive()) {
             while (opModeIsActive()) {
                 while (runtime.seconds() < 3) {
+                    bot.target = 0;
                     bot.closeClaw();
-                    bot.odometry.followTrajectory(findBlueBProp);
+                    //bot.odometry.followTrajectory(findBlueFProp);
 
                     if (bot.pipeline.isPropRight()){
                         spikeLoc = SPIKE_LOC.RIGHT;
@@ -172,21 +185,21 @@ public class AutoBB_WithCamera extends LinearOpMode {
                     case RIGHT:
                         telemetry.addLine("RIGHT SPIKE");
                         if (!bot.odometry.isBusy()) {
-                            bot.odometry.followTrajectorySequenceAsync(BBdriveToRSpike);
+                            bot.odometry.followTrajectorySequenceAsync(BFdriveToRSpike);
                             spikeLoc = SPIKE_LOC.IDLE;
                             break;
                         }
                     case CENTER:
                         telemetry.addLine("CENTER SPIKE");
                         if (!bot.odometry.isBusy()) {
-                            bot.odometry.followTrajectorySequenceAsync(BBdriveToCSpike);
+                            bot.odometry.followTrajectorySequenceAsync(BFdriveToCSpike);
                             spikeLoc = SPIKE_LOC.IDLE;
                             break;
                         }
                     case LEFT:
                         telemetry.addLine("LEFT SPIKE");
                         if (!bot.odometry.isBusy()) {
-                            bot.odometry.followTrajectorySequenceAsync(BBdriveToLSpike);
+                            bot.odometry.followTrajectorySequenceAsync(BFdriveToLSpike);
                             spikeLoc = SPIKE_LOC.IDLE;
                             break;
                         }
@@ -199,5 +212,6 @@ public class AutoBB_WithCamera extends LinearOpMode {
                 telemetry.update();
             }
         }
+
     }
 }
